Move inside ui.map.gms package
