+++ ground/src/main/java/com/google/android/ground/system/GeocodingManager.kt
@@ -18,10 +18,10 @@ package com.google.android.ground.system
 import android.content.res.Resources
 import android.location.Geocoder
 import com.google.android.ground.R
-import com.google.android.ground.exts.GmsExt.center
 import com.google.android.ground.rx.Schedulers
 import com.google.android.ground.rx.annotations.Cold
 import com.google.android.ground.ui.map.Bounds
+import com.google.android.ground.ui.map.gms.GmsExt.center
 import io.reactivex.Single
 import java.io.IOException
 import java8.util.Optional
