+++ ground/src/main/java/com/google/android/ground/ui/map/gms/GmsExt.kt
@@ -1,12 +1,9 @@
-package com.google.android.ground.exts
+package com.google.android.ground.ui.map.gms
 
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.ground.model.geometry.Coordinate
 import com.google.android.ground.model.geometry.Geometry
 import com.google.android.ground.ui.map.Bounds
-import com.google.android.ground.ui.map.gms.toGoogleMapsObject
-import com.google.android.ground.ui.map.gms.toLatLng
-import com.google.android.ground.ui.map.gms.toModelObject
 
 /** Extensions for indirectly using GMS functions in map-provider agnostic codebase. */
 object GmsExt {
